+++ src/bittytax/conv/mergers/__init__.py
@@ -1,1 +1,1 @@
-from . import bscscan, etherscan, hecoinfo, polygonscan, snowtrace
+from . import bscscan, coinbase, etherscan, hecoinfo, polygonscan, snowtrace